prevent empty data payload from being send to host (#5)
* adds per host verbosity setting

This commit allows for a per host verbosity setting to control the
amount of debug output.  The device can now set `itential_verbosity` to
any value in the range of 1 to 5.  If the `itential_verbosity` settings
is less than the `ansible_verbosity` setting, the `ansible_verbosity`
setting will take precedence.

* prevent empty data payload from being send to host

This commit addresses an issue where an empty data payload would be send
to the host as an empty array or empty object.  This change will now
check if the data is empty and not attempting to send it to the host if
it is.

* clean up lint errors

* update workflow yaml lint error