remove data encoding to json (#9)
This commit removes the logic to encode the payload before sending the
request.  This library should not attempt to infer the data type and
simply add the data to the request exactly as it was passed to the
function.  It is the responsibility of the upstream function to call the
function with the appropriately encoded data.

This commit will now check if the value of data is type of bytes and
raise an exception if it is not